@@ -1,21 +1,29 @@
 {
   "name": "xpr-dashboard",
-  "version": "3.0.1",
+  "version": "4.0.0",
   "description": "Experiments Dashboard for Frontier",
-  "main": "./lib/app.js",
+  "main": "./index.js",
   "author": "Dan Crews <crewsd@gmail.com>",
+  "homepage": "https://github.com/XPRMNTL/xpr-dashboard",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/XPRMNTL/xpr-dashboard"
+  },
+  "bugs": {
+    "url": "https://github.com/XPRMNTL/xpr-dashboard/issues"
+  },
   "dependencies": {
-    "bluebird": "^2.9.25",
-    "connect-redis": "^2.3.0",
-    "debug": "^2.1.3",
-    "glob": "^5.0.5",
-    "jade": "^1.9.2",
-    "node-uuid": "^1.4.3",
-    "passport": "^0.2.1",
+    "babel": "^5.6.4",
+    "bluebird": "^2.9.30",
+    "debug": "^2.2.0",
+    "express-session": "^1.11.3",
+    "jack-stack": "^1.0.0",
+    "jack-stack-redis": "^1.1.0",
+    "jack-stack-stylus": "^1.1.0",
+    "jade": "^1.11.0",
+    "passport": "^0.2.2",
     "passport-github": "^0.1.5",
-    "redis": "^0.12.1",
-    "simple-stack-common": "^0.2.3",
-    "stylus": "^0.51.1",
+    "redis-url": "^1.2.0",
     "superagent": "^1.2.0",
     "superagent-defaults": "^0.1.12",
     "xpr-dash-mongodb": "^2.0.0",
@@ -30,6 +38,6 @@
     "test": "NODE_ENV=test `pwd`/node_modules/.bin/mocha --reporter spec test/server/**/*"
   },
   "engines": {
-    "node": "0.10.x"
+    "node": "0.12.x"
   }
 }
