@@ -1,22 +1,34 @@
-(function() {
-  'use strict';
+'use strict';
 
-  var superagent = require('superagent-defaults')()
-    , debug = require('debug')('feature:lib:github')
-    , Promise = Promise || require('bluebird')
-  ;
+/**
+ * Module Dependencies
+ */
+import superagentDefaults from 'superagent-defaults';
+import debugLib from 'debug';
+import Promise from 'bluebird';
 
-  module.exports = GithubAPI;
+/**
+ * Local Constants
+ */
+const superagent = superagentDefaults();
+const debug = debugLib('xpr:lib:github');
 
-  function GithubAPI(user, token) {
+
+/**
+ * Export the Class
+ */
+export default class GithubAPI {
+
+
+  constructor(user, token) {
     this.user = user;
 
-    this.request = superagent.on('request', function(request) {
+    this.request = superagent.on('request', (request) => {
       request.set('Accept', 'application/vnd.github.raw+json');
       if (token) request.auth(token, 'x-oauth-basic');
 
-      request.setCb = function(cb, dfd) {
-        var fail = function(err) {
+      request.setCb = (cb, dfd) => {
+        var fail = (err) => {
           if (typeof err === 'string') {
             err = new Error(err);
           }
@@ -24,12 +36,12 @@
           if (cb) cb(err);
           if (dfd) dfd.reject(err);
         };
-        var win = function(resp) {
+        var win = (resp) => {
           if (cb) cb(null, resp.body, resp.headers);
           if (dfd) dfd.resolve(resp.body);
         };
         request.on('error', fail);
-        request.on('response', function(resp) {
+        request.on('response', (resp) => {
           if (resp.ok) return win(resp);
 
           debug('Request failed: ' + resp.status);
@@ -44,46 +56,47 @@
     return this;
   }
 
-  GithubAPI.prototype.fetchOrgs = function(cb) {
-    var url = 'https://api.github.com/user/orgs'
-      , self = this;
 
-    return new Promise(function(resolve, reject) {
-      self.request
+  fetchOrgs(cb) {
+    let url = 'https://api.github.com/user/orgs';
+
+    return new Promise((resolve, reject) => {
+      this.request
         .get(url)
         .setCb(cb, { resolve: resolve, reject: reject })
         .end();
     });
-  };
+  }
 
-  GithubAPI.prototype.filterOrgs = function(orgs) {
-    var okOrgs = this.orgs
-      , isMember = false;
-    orgs.map(function(org) {
+
+  filterOrgs(orgs) {
+    let okOrgs = this.orgs;
+    let isMember = false;
+
+    orgs.map((org) => {
       if (~ okOrgs.indexOf(org.login.toLowerCase())) isMember = true;
     });
 
     if (! isMember) throw { message: 'Not a Member of required Orgs', name: 'NOT_AUTHORIZED' };
-  };
+  }
 
-  GithubAPI.prototype.fetchRepos = function() {
-    var self = this;
 
-    return new Promise(function(resolve) {
-      var appMap = {}
-        , _request = self.request;
+  fetchRepos() {
+    return new Promise((resolve) => {
+      let appMap = {};
+      let _request = this.request;
 
-      var promises = self.orgs.map(function(orgName) {
-        return new Promise(function(resolve) {
-          fetchPage(1, function(err, data, headers) {
-            var links = headers.link
-              , _promises = []
-              , matches, pages;
+      let promises = this.orgs.map((orgName) => {
+        return new Promise((resolve) => {
+          fetchPage(1, (err, data, headers) => {
+            let links = headers.link;
+            let _promises = [];
+            let matches, pages;
 
             if (links) {
               matches = links.match(/\?page=[^?]*\?page=(\d+)/);
               pages = parseInt(matches[1]);
-              for (var i=2; i<=pages; i++) {
+              for (let i=2; i<=pages; i++) {
                 _promises.push(fetchPage(i));
               }
             }
@@ -96,8 +109,8 @@
           }
 
           function fetchPage(page, cb) {
-            return new Promise(function(resolve, reject) {
-              var url = 'https://api.github.com/orgs/' + orgName + '/repos?page=' + page + '&per_page=100';
+            return new Promise((resolve, reject) => {
+              var url = `https://api.github.com/orgs/${orgName}/repos?page=${page}&per_page=100`;
 
               _request.get(url)
                 .setCb(function(err, data, headers) {
@@ -115,6 +128,6 @@
           resolve(appMap);
         });
     });
-  };
+  }
 
-})();
+}
